backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
anthropic==0.7.8
openai==1.3.5
langchain==0.0.350
langchain-anthropic==0.0.1
langchain-openai==0.0.2
langsmith==0.0.66

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Utilities
httpx==0.25.2
python-multipart==0.0.6
tiktoken==0.5.2

